+++ app/src/main/java/com/benio/demoproject/progress/PercentFormatter.java
@@ -3,12 +3,12 @@ package com.benio.demoproject.progress;
 import android.support.annotation.NonNull;
 
 public class PercentFormatter implements Formatter {
-    private int oldValue = -1;
+    private float oldValue = -1;
     private String oldStr = null;
 
     @NonNull
     @Override
-    public String format(int value) {
+    public String format(float value) {
         if (oldValue != value) {
             oldValue = value;
             oldStr = value + "%";
